requirements.txt
langchain_google_vertexai == 2.0.27
langchain ==  0.3.26
google-cloud-speech == 2.33.0
google-cloud-translate == 3.21.1
pydantic == 2.11.7
dotenv
langgraph
beautifulsoup4 
google-cloud-firestore
sentence_transformers == 5.0.0
aiohttp
langgraph_supervisor == 0.0.27
langchain_google_genai == 2.1.8
fastapi
requests 
rasterio 
matplotlib 
geopandas 
folium